refactor(api): TYS-180 simplify API by merging read/write routers into unified data router
The split read/write router architecture caused 405 Method Not Allowed
errors because Django Ninja creates separate URL patterns per Router
instance mounted at the same prefix. Django's URL resolver matched only
the first pattern, rejecting POST/PATCH/DELETE registered on the second.

- Merge blog/api/read/ and blog/api/write/ into blog/api/data/
- Fix unthrottled /register/ and /profile/ endpoints (security)
- Add missing CSRF protection on /resend-verification/
- Add JsonMethodNotAllowedMiddleware for JSON 405 on API paths
- Content-negotiate CSRF_FAILURE_VIEW (JSON for API, HTML for admin)
- Extract shared _request_data helpers to blog/api/utils.py
- Update all test imports and assertions

Closes #56

Made-with: Cursor
