+++ tests/unit/test_ninja_auth_preview.py
@@ -3,32 +3,24 @@ from django.test import Client
 
 
 @pytest.mark.django_db
-class TestNinjaAuthPreview:
-    def test_preview_openapi_is_available(self, client):
-        resp = client.get("/api/_ninja/auth/openapi.json")
-
-        assert resp.status_code == 200
-        data = resp.json()
-        assert data["openapi"].startswith("3.")
-        assert data["info"]["title"] == "claude_rest_api auth preview"
-
-    def test_csrf_preview_returns_token_and_cookie(self, client):
-        resp = client.get("/api/_ninja/auth/csrf/")
+class TestNinjaAuth:
+    def test_csrf_returns_token_and_cookie(self, client):
+        resp = client.get("/api/auth/csrf/")
 
         assert resp.status_code == 200
         assert resp.json()["csrfToken"]
         assert resp.cookies["csrftoken"].value
 
-    def test_current_user_preview_requires_authentication(self, client):
-        resp = client.get("/api/_ninja/auth/user/")
+    def test_current_user_requires_authentication(self, client):
+        resp = client.get("/api/auth/user/")
 
         assert resp.status_code == 403
         assert resp.json()["detail"] == "Authentication credentials were not provided."
 
-    def test_current_user_preview_returns_current_user_payload(self, client, user):
+    def test_current_user_returns_current_user_payload(self, client, user):
         client.force_login(user)
 
-        resp = client.get("/api/_ninja/auth/user/")
+        resp = client.get("/api/auth/user/")
 
         assert resp.status_code == 200
         data = resp.json()
@@ -37,26 +29,26 @@ class TestNinjaAuthPreview:
         assert data["profile"]["role"] == "user"
         assert data["can_manage_tags"] is False
 
-    def test_logout_preview_requires_authentication(self, client):
-        resp = client.post("/api/_ninja/auth/logout/")
+    def test_logout_requires_authentication(self, client):
+        resp = client.post("/api/auth/logout/")
 
         assert resp.status_code == 403
         assert resp.json()["detail"] == "Authentication credentials were not provided."
 
-    def test_logout_preview_succeeds_with_valid_csrf(self, user):
+    def test_logout_succeeds_with_valid_csrf(self, user):
         client = Client(enforce_csrf_checks=True)
         client.force_login(user)
 
-        csrf_resp = client.get("/api/_ninja/auth/csrf/")
+        csrf_resp = client.get("/api/auth/csrf/")
         csrf_token = csrf_resp.cookies["csrftoken"].value
 
         resp = client.post(
-            "/api/_ninja/auth/logout/",
+            "/api/auth/logout/",
             HTTP_X_CSRFTOKEN=csrf_token,
         )
 
         assert resp.status_code == 200
         assert resp.json()["detail"] == "Logged out."
 
-        after = client.get("/api/_ninja/auth/user/")
+        after = client.get("/api/auth/user/")
         assert after.status_code == 403
